build: :rocket: deploy finance and frendlend (#67)
